feat: Implement Choose from Gallery Action (#63)
Update the iOS library.
Add the `chooseFromGallery` action to make a bridge to the native code. The action call is currently being made in `takePicture` so no breaking changes are added to Android's implementation.
Add the Info.plist entry to prevent the automatic presentation of the Limited Selection Access. This presentation is controlled by the plugin itself.